+++ e-filesDownload/src/main/java/eu/unifiedviews/plugins/extractor/filesdownload/FilesDownload.java
@@ -3,7 +3,12 @@ package eu.unifiedviews.plugins.extractor.filesdownload;
 import java.io.IOException;
 import java.net.URI;
 import java.text.NumberFormat;
+import java.util.ArrayList;
 import java.util.Date;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
 
 import org.apache.commons.codec.digest.DigestUtils;
 import org.apache.commons.httpclient.params.HttpConnectionManagerParams;
@@ -20,6 +25,7 @@ import org.apache.commons.vfs2.auth.StaticUserAuthenticator;
 import org.apache.commons.vfs2.cache.NullFilesCache;
 import org.apache.commons.vfs2.impl.DefaultFileSystemConfigBuilder;
 import org.apache.commons.vfs2.impl.StandardFileSystemManager;
+import org.apache.commons.vfs2.provider.UriParser;
 import org.apache.commons.vfs2.provider.ftp.FtpFileSystemConfigBuilder;
 import org.apache.commons.vfs2.provider.ftps.FtpsFileSystemConfigBuilder;
 import org.apache.commons.vfs2.provider.sftp.SftpFileSystemConfigBuilder;
@@ -55,6 +61,8 @@ import eu.unifiedviews.plugins.extractor.httpdownload.HttpDownloadConfig_V1;
 public class FilesDownload extends AbstractDpu<FilesDownloadConfig_V1> {
 
     private static final Logger LOG = LoggerFactory.getLogger(FilesDownload.class);
+
+    public static final String SUPPORTED_PROTOCOLS = "dpu.uv-e-filesDownload.allowed.protocols";
 
     @RdfConfiguration.ContainsConfiguration
     @DataUnit.AsInput(name = "config", optional = true)
@@ -99,6 +107,8 @@ public class FilesDownload extends AbstractDpu<FilesDownloadConfig_V1> {
 
         if (config.isIgnoreTlsErrors()) {
             Protocol.registerProtocol("https", new Protocol("https", (ProtocolSocketFactory) new EasySSL(), 443));
+        } else {
+            Protocol.registerProtocol("https", new Protocol("https", (ProtocolSocketFactory) new TLS12(), 443));
         }
 
         final NumberFormat numberFormat = NumberFormat.getNumberInstance();
@@ -117,6 +127,14 @@ public class FilesDownload extends AbstractDpu<FilesDownloadConfig_V1> {
             if (ctx.canceled()) {
                 throw ContextUtils.dpuExceptionCancelled(ctx);
             }
+
+            if (!checkURIProtocolSupported(vfsFile.getUri())) {
+                ContextUtils.sendWarn(this.ctx, "FilesDownload.protocol.not.supported", "FilesDownload.protocol.not.supported.long",
+                        vfsFile.getUri(),
+                        getSupportedProtocols());
+                continue;
+            }
+
             // If user name is not blank, then we prepare for autentification.
             if (StringUtils.isNotBlank(vfsFile.getUsername())) {
                 final StaticUserAuthenticator staticUserAuthenticator;
@@ -203,6 +221,43 @@ public class FilesDownload extends AbstractDpu<FilesDownloadConfig_V1> {
                 }
             }
         }
+    }
+
+    private boolean checkURIProtocolSupported(String uri) {
+        Map<String, String> environment = this.ctx.getExecMasterContext().getDpuContext().getEnvironment();
+        String supportedProtocols = environment.get(SUPPORTED_PROTOCOLS);
+        if (StringUtils.isEmpty(supportedProtocols)) {
+            return true;
+        }
+
+        final String scheme = UriParser.extractScheme(uri);
+        String[] supportedSchemes = supportedProtocols.trim().split(",");
+        Set<String> supportedSet = new HashSet<>();
+        for (String s : supportedSchemes) {
+            supportedSet.add(s);
+        }
+
+        if (StringUtils.isEmpty(scheme) && !supportedSet.contains("file")) {
+            return false;
+        }
+
+        return supportedSet.contains(scheme);
+    }
+
+    private List<String> getSupportedProtocols() {
+        Map<String, String> environment = this.ctx.getExecMasterContext().getDpuContext().getEnvironment();
+        String supportedProtocols = environment.get(FilesDownload.SUPPORTED_PROTOCOLS);
+        if (StringUtils.isEmpty(supportedProtocols)) {
+            return null;
+        }
+
+        String[] supportedProtocolsArray = supportedProtocols.trim().split(",");
+        List<String> protocols = new ArrayList<>();
+        for (String protocol : supportedProtocolsArray) {
+            protocols.add(protocol);
+        }
+
+        return protocols;
     }
 
 }