+++ e-relationalFromSql/CHANGELOG.md
@@ -1,7 +1,16 @@
 E-RelationalFromSql
 ----------
 
-v2.1.1-SNAPSHOT
+v2.1.3-SNAPSHOT
+---
+* Update to helpers 2.1.3
+
+v2.1.2
+---
+* Fixed documentation (About)
+* Update to API 2.1.2 (properly displayed about tab in UTF-8 encoding)
+
+v2.1.1
 ---
 * Building with Oracle support is now optional via Maven profiles
 * Improved description (About)